+++ app/src/main/java/info/nightscout/androidaps/plugins/PumpDanaR/comm/MessageHashTable.java
@@ -56,11 +56,11 @@ public class MessageHashTable {
             put(new MsgSettingGlucose());            // 0x3209 CMD_SETTING_V_GLUCOSEandEASY
             put(new MsgSettingPumpTime());           // 0x320A CMD_SETTING_V_TIME_I
             put(new MsgSettingUserOptions());        // 0x320B CMD_SETTING_V_USER_OPTIONS
-            put(new MsgGetUserOptions());            // 0x320B CMD_SETTING_V_USER_OPTIONS - trying to get uptions
             put(new MsgSettingActiveProfile());      // 0x320C CMD_SETTING_V_PROFILE_NUMBER
             put(new MsgSettingProfileRatiosAll());   // 0x320D CMD_SETTING_V_CIR_CF_VALUE
             put(new MsgSetSingleBasalProfile());     // 0x3302 CMD_SETTING_BASAL_INS_S
             put(new MsgSetBasalProfile());           // 0x3306 CMD_SETTING_BASAL_PROFILE_S
+            put(new MsgSetUserOptions());            // 0x330B CMD_SETTING_USER_OPTIONS_S
             put(new MsgSetActivateBasalProfile());   // 0x330C CMD_SETTING_PROFILE_NUMBER_S
             put(new MsgHistoryAllDone());            // 0x41F1 CMD_HISTORY_ALL_DONE
             put(new MsgHistoryAll());                // 0x41F2 CMD_HISTORY_ALL
